Colocate Graphql Queries
