--- packages/frontend-airdos/src/pages/LoginPage/LoginPage.queries.tsx
@@ -1,9 +1,0 @@
-import gql from 'graphql-tag'
-
-const VERIFY_USER = gql`
-  query verifyUser($username: String!, $password: String!) {
-    verifyUser(username: $username, password: $password)
-  }
-`
-
-export { VERIFY_USER }
